Move most code from falco_init() to individual app actions
Each file below app_actions/ exports a function that creates an action
object. The run function (and when needed, deinit function) are static
functions in each file.

Any state that needs to be shared betweeen actions, or between the run
and deinit functions, resides in app::state(), so the moved code stays
pretty much as-is, other than replacing stack variables with member
variables in app_state.

Signed-off-by: Mark Stemm <mark.stemm@gmail.com>